move permutations to core/symgroup
