+++ topos/core/__init__.py
@@ -1,4 +1,5 @@
-from .shape import Shape
+from .shape     import Shape
+from .symgroup  import SymGroup
 #old
 #from .field import Field
 #from .functional import Functional, GradedFunctional
